Search: replace the URL to match current location
The algolia items are stored using the production documentation. For
development, this is quite a pain because we can't use the search
otherwise it will redirect us to the production documentation.

Instead, we replace the suggestion href domain / port / protocol with
our current location domain / port / protocol. And if we are in
development mode, then we strip the `/doc` part.